core: Add a config flag for enabling BasicManagement
Remove the coupling to MediaRenderer: allow BasicManagement on any
rygel plugin, provided the configuration specifies that diagnostics
should be started.

Add a "+DIAGE+" capability that is exposed in the device description
when diagnostics is enabled.

This change means that by default the service will not run. Add this to
rygel.conf to enable:
  [Playbin]
  diagnostics=true